Update.
There is an error: ID3D11DeviceContext::DrawIndexed: A Vertex Shader is always required when drawing, but none is currently bound.

ID3D11DeviceContext::DrawIndexed: Rasterization Unit is enabled (PixelShader is not NULL or Depth/Stencil test is enabled and RasterizedStream is not D3D11_SO_NO_RASTERIZED_STREAM) but position is not provided by the last shader before the Rasterization Unit.