Added more lights and lighting calculation.
I added a point light and a spotlight light to the light array with different added option values.

Now each of the light in the scene have their diffuse and specular reflection color on a common ambient surface calculated and added to each light in the scene, by using a for loop and a switch statement.