Update Attenuate.
I attenuate the light depending on the distance between the pixel surface world position and the position of the light.

If the pixel world position is beyond light range, no light shined on the pixel else the light shining is reduced the farther the pixel is away from the light position.